streamlet/requirements.txt
streamlit
python-dotenv
requests
scikit-learn
matplotlib
pandas
pinecone-client
sentence-transformers
reportlab
